# src/user_service/openapi.yml
openapi: 3.0.3
info:
  version: 1.0.0
  title: User Service API
paths:
  /user/register:
    post:
      summary: Register a new user
      operationId: registerUser
      requestBody:
        description: A JSON object containing username and password
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: User successfully registered
        '400':
          description: Bad Request
        '409':
          description: Username already exists
  /user/login:
    post:
      summary: Log user into the system
      operationId: loginUser
      requestBody:
        description: A JSON object containing username and password
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: User successfully authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationToken'
        '403':
          description: Incorrect username or password
  /user/update:
    put:
      security:
        - bearerAuth: []
      summary: Update user information
      operationId: updateUser
      requestBody:
        description: A JSON object containing user information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInfo'
        required: true
      responses:
        '200':
          description: User information successfully updated
        '400':
          description: Bad Request
        '401':
          description: User unauthorized
        '404':
          description: User not found
  /post:
    post:
      security:
        - bearerAuth: []
      summary: Create post
      operationId: createPost
      requestBody:
        description: Post content
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
              required:
                - content
        required: true
      responses:
        '200':
          description: Post successfully created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                required:
                  - id
        '400':
          description: Bad Request
        '401':
          description: User unauthorized
        '404':
          description: User not found
  /post/{id}:
    put:
      security:
        - bearerAuth: []
      summary: Update post
      operationId: updatePost
      parameters:
        - name: id
          in: path
          description: Post id
          required: true
          schema:
            type: integer
      requestBody:
        description: Post id and new content
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
              required:
                - content
        required: true
      responses:
        '200':
          description: Post successfully updated
        '400':
          description: Bad Request
        '401':
          description: User unauthorized
        '404':
          description: User not found
    delete:
      security:
        - bearerAuth: []
      summary: Delete post
      operationId: deletePost
      parameters:
        - name: id
          in: path
          description: Post id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Post successfully deleted
        '400':
          description: Bad Request
        '401':
          description: User unauthorized
        '404':
          description: User not found
    get:
      security:
        - bearerAuth: []
      summary: Get post
      operationId: getPost
      parameters:
        - name: id
          in: path
          description: Post id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Bad Request
        '401':
          description: User unauthorized
        '404':
          description: User not found
  /posts:
    get:
      security:
        - bearerAuth: []
      summary: List posts
      operationId: listPosts
      parameters:
        - name: limit
          in: query
          description: Limit of posts
          required: true
          schema:
            type: integer
        - name: offset
          in: query
          description: Offset of posts
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '400':
          description: Bad Request
        '401':
          description: User unauthorized
        '404':
          description: User not found
  /like:
    post:
      security:
        - bearerAuth: []
      summary: Like post
      operationId: likePost
      parameters:
        - name: post
          in: query
          description: Post id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Post successfully liked
        '400':
          description: Bad Request
        '401':
          description: User unauthorized
        '404':
          description: User not found
  /view:
    post:
      security:
        - bearerAuth: []
      summary: View post
      operationId: viewPost
      parameters:
        - name: post
          in: query
          description: Post id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Post successfully viewed
        '400':
          description: Bad Request
        '401':
          description: User unauthorized
        '404':
          description: User not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      required:
        - username
        - password
      type: object
      properties:
        username: 
          type: string
        password: 
          type: string
    UserInfo:
      type: object
      properties:
        firstName: 
          type: string
        lastName: 
          type: string
        dateOfBirth: 
          type: string
          format: date
        mail: 
          type: string
          format: email
        phone: 
          type: string
    AuthenticationToken:
      type: object
      properties:
        token: 
          type: string
    Post:
      required:
        - id
        - username
        - content
      type: object
      properties:
        id: 
          type: integer
        username: 
          type: string
        content: 
          type: string